% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level2summarize.R
\name{level2summarize}
\alias{level2summarize}
\title{Use dplyr to calculate level 2 statistics}
\usage{
level2summarize(data, a, b)
}
\arguments{
\item{data}{A data frame, or a data frame extension (e.g. a tibble).}

\item{a}{The name of the column from the data frame to group by.}

\item{b}{The name of the column from the data frame to be summarized.}
}
\value{
A summary table with sample size, number of missing value, mean, median, standard deviation, minimum, and maximum for the given column b grouped by the column a.
}
\description{
This function is a wrapper for the \code{\link[dplyr:summarise]{dplyr::summarize()}} function from the dplyr package. The function is fortified to handle the missing values (NA), and it can provide a summary table with many descriptive statistics for a variable grouped by a specify categorical variable. In this case, this function can be used in multilevel models, and provide the statistics for the level 2 units (e.g. the groups).
}
\examples{
# 1. A real dataset example
level2summarize(mtcars, am, mpg) # summarize mpg based on am

# 2. A fake datseat example to show parameter a can be numeric or factor
a <- as.factor(c(rep(1, 10), rep(2, 10)))
b <- as.numeric(c(rep(1, 10), rep(-1, 10))) # a as numeric
data1 <- data.frame(a, b)
level2summarize(data1, a, b)
a <- as.numeric(c(rep(1, 10), rep(2, 10)))
b <- as.numeric(c(rep(1, 10), rep(-1, 10))) # a as factor
data1 <- data.frame(a, b)
level2summarize(data1, a, b)




}
\seealso{
\code{\link[dplyr:summarise]{dplyr::summarize()}}
}
